Percona mysql cluster support
If have set up a percona mysqli cluster and was having trouble load balancing the thing on qcubed. You can probably do this with PDO as well but I decided it was simpler to load balance it in qcubed. 

In configuration.inc.php you simply add something like this:
define('DB_CONNECTION_1', serialize(array(
				'adapter'       => 'MySqli5Cluster',
				'server'		=> array('mysql1','mysql2','mysql3'),

For installation instructions on percona mysql cluster see: 

http://www.percona.com/doc/percona-xtradb-cluster/5.6/installation.html