Refactoring sending of commands to javascript widgets. This work paves the way to eventually removing the eval in qcubed.js. This is an intermediate step in the process, but one that is a good milestone for testing.
Fixed a couple of things in the Examples code during testing.

Added an independent test of the new javascript file injection mechanism. The new mechanism allows injected files to be loaded in parallel, and for some ajax processing to occur while the load is happening as well, overall improving ajax performance in these special situations.