Update QCache.class.php
QCache runs sprintf multiple times in the same method. I've changed this so the variable $filepath is used.

Also, it sometimes happens that an unlink fails because a file does not exist. 
This should be impossible since we are already checking if the file_exists at the start of the method. But the reality is different. 
My guess is that there are multiple GetData calls by different users and that they sometimes overlap. 

To fix this I added an @ to the unlink. This will avoid a crash if the file does not exist.