Making Minimize into a QApplication global instead of a define so that controls can individually turn it on and off as needed.
The primary thing I am fixing is that in certain situations, like when html objects are styled with inline-block styling, the tabs and returns that are inserted to make the html look good will also add a 4px wide space between objects, causing different drawing when Minimize is on and off. This allows a particular control to always draw minimized so that doesn't happen.