Refactor of the QQuery code. Does the following: - Documenting functions - Remove most warnings - Removes the QQBaseNode. All nodes derive from QQNode. - Creates the concept of a QColumnNode and QTableNode to differentiate from leaf and stem nodes. This allows PHPs own type checking system to do some type checking. - Moves Fields to the nodes to make them more available to the join code. Slight speed improvement as a result, but it at least avoids the need to do a function call on a variable named class. - GetColumnAlias is not overloaded to also do the join of nodes. They are separate functions for better understanding, and so that GetColumnAlias will only apply to column nodes and not table nodes.
This will require a code regeneration, but otherwise, should not impact existing code. Their may be situations where someone is using a table node in a comparison. This is no longer allowed, you must specify the primary key of the table from now on, or some other column node.

Did not remove the HTML generating code. For another day, or someone else to do.