Don't use output buffering when on cli
We usually run all qcubed applications in a webbrowser. But sometimes I'd like to write scripts with qcubed that I run on the command line. 
When executing a script on the command line, it makes little sense to buffer the output. You actually want the output straight away. 

I don't know if this change belongs in the core of qcubed, but I can't see how to change this in another manner.