Fix to a timing problem with checkboxes. This issue was probably the original reason for a checkbablecontrols hidden input and handler.
The problem is that for checkboxes and radio buttons, a "click" event is equivalent to a "change" event for our purposes. We want to be able to attach a click event to a checkbox, and during the action, see that the checkbox has changed. This was not happening for ajax handlers because the click would fire before the change, and so we would not post the change until the next ajax event.

This fix detects clicks on checkboxes and radio buttons, and then registers them as changed so that we send the value of the checkbox or radio button to the server along with any other event handlers.