Method created
GetRandomString method in QString is new and facilitates the feature

It returns a string of a given length, the characters of which are extracted
from the character set supplied as the second argument and are in random
order. In case a character set is not supplied, alphanumeric character set
is used by default.