Many dialog improvements that have other improvements.
This fix allows us to finally get to the point where you can bring up a dialog on an ad-hoc basis without giving the dialog a parent control at all or saving the dialog in the form state long-term.

Dialogs now default to being modal, since that is the most common for dialogs.

Creating a new class of javascript events that are "extra-low". These events will be executed only after the ajaxq is completely empty. So any new events that are added to the queue in response to an ajax event will further delay the execution of the final event. This is currently useful for closing dialogs so that we can process the close of a dialog after all other events have been processed. This will also be helpful in the future as part of an automated test tool, which will be used to detect the end of a chain of ajax events.