Cleaning up the QDbBackedSessionHandler
I needed this at work so I cleaned it up a bit. It was written in 2012 and was missing some documentation.
I added an example SQL CREATE statement for the database which makes it easier for users to get started with it.
I also removed the GetTables() in the Initialize because I don't think it's a good idea to do this every single request. It will crash anyway, and a database table won't just disappear...

I also moved the concatenation of the session_name with the session_id to a seperate function.