InsertOrUpdate now returns the autoincremented ID
InsertOrUpdate is a great method that is currently only used by the dbsessionhandler. It allows you to remove a query to determine if you want to insert or update because mysql will do this for you. 
Unfortunately, when you use this method, you didn't get the LAST_INSERT_ID on insert. I added this, plus when there is a duplicate, qcubed will return the id found in the update. 

To simplify the INSERT statement I also used the alternative INSERT syntax. INSERT INTO tablename SET col=val in stead of INSERT INTO tablename (col) values(val)

I'm running dbbackedsessionhandler in production.and it sees no difference between the old and new function. So there shouldn't be an impact on the rest of qcubed.