$strFormId is useless
$strFormId is useless. I think @spekary already tried to refactor some of it. 
With this commit I try to remove strFormId and pass the object itself. 

A possible use case is requesting a control by strControlId in a static method call. This is currently impossible, but after this commit you have the objFormId and can actually do this. 
I have also added a __toString() method in an attempt to save backwards compatibility. If someone does use the strFormId, it gets string cast automatically.